app/requirements.txt
streamlit>=1.31.0
gspread>=6.0.0
google-auth>=2.27.0
pandas>=2.0.0
reportlab>=4.0.0
